move function `load_sys_from_str` into seperate module
